39 done

+++ src/main/java/com/spring/community/community/dto/QuestionDTO.java
@@ -11,8 +11,8 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Integer id;
-    private Integer creator;
+    private Long id;
+    private Long creator;
     private String title;
     private String description;
     private String tag;
